opensky_tracker/requirements.txt
trino
pandas
python-dotenv
